+++ conjur_api/providers/simple_credentials_provider.py
@@ -3,8 +3,9 @@ SimpleCredentialsProvider module
 
 This module holds the SimpleCredentialsProvider class
 """
-from typing import Dict
 from copy import deepcopy
+from typing import Dict
+
 from conjur_api.interface import CredentialsProviderInterface
 from conjur_api.models import CredentialsData
 
@@ -15,6 +16,7 @@ class SimpleCredentialsProvider(CredentialsProviderInterface):
 
     This class implement a naive, in-memory implementation of the CredentialsProvider Interface
     """
+
     def __init__(self):
         self._credentials: Dict[str, CredentialsData] = {}