+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [8.0.0] - 2022-05-25
 
+### Added
+- The `get_server_info` method is now available in SDK. It is only supported against Conjur enterprise server
+- Authentication via OIDC is now supported
+- Supplying api token manually is now supported 
 ### Changed
 - Abstract authentication flow into new `AuthenticationStrategyInterface`
   [conjur-api-python#20](https://github.com/cyberark/conjur-api-python/pull/20)